Added exporting report

+++ SchoolAppCoreMVC/Controllers/StudentsController.cs
@@ -6,51 +6,53 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SchoolAppCoreMVC.Models;
+using Microsoft.Reporting.NETCore;
 
 namespace SchoolAppCoreMVC.Controllers
 {
     public class StudentsController : Controller
     {
         private readonly SchoolContext _context;
+        private readonly IWebHostEnvironment _webHostEnvironment;
 
-        public StudentsController(SchoolContext context)
+        public StudentsController(SchoolContext context, IWebHostEnvironment webHostEnvironment)
         {
             _context = context;
+            _webHostEnvironment = webHostEnvironment;
         }
 
-        // GET: Students
         public async Task<IActionResult> Index()
         {
             return View(await _context.Students.ToListAsync());
         }
 
-        // GET: Students/Details/5
+        public IActionResult ExportToPdf()
+        {
+            var students = _context.Students.ToList();
+
+            string reportPath = Path.Combine(_webHostEnvironment.ContentRootPath, "StudentReport.rdlc");
+
+            LocalReport report = new LocalReport();
+            report.ReportPath = reportPath;
+
+            report.DataSources.Add(new ReportDataSource("dsStudents", students));
+
+          
+            var pdf = report.Render("PDF");
+
+            return File(pdf, "application/pdf", "StudentReport.pdf");
+        }
+
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var student = await _context.Students
-                .FirstOrDefaultAsync(m => m.StudentID == id);
-            if (student == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var student = await _context.Students.FirstOrDefaultAsync(m => m.StudentID == id);
+            if (student == null) return NotFound();
             return View(student);
         }
 
-        // GET: Students/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        public IActionResult Create() => View();
 
-        // POST: Students/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("StudentID,FirstName,LastName,EnrollmentDate")] Student student)
@@ -64,34 +66,19 @@ namespace SchoolAppCoreMVC.Controllers
             return View(student);
         }
 
-        // GET: Students/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var student = await _context.Students.FindAsync(id);
-            if (student == null)
-            {
-                return NotFound();
-            }
+            if (student == null) return NotFound();
             return View(student);
         }
 
-        // POST: Students/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("StudentID,FirstName,LastName,EnrollmentDate")] Student student)
         {
-            if (id != student.StudentID)
-            {
-                return NotFound();
-            }
-
+            if (id != student.StudentID) return NotFound();
             if (ModelState.IsValid)
             {
                 try
@@ -101,56 +88,32 @@ namespace SchoolAppCoreMVC.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!StudentExists(student.StudentID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!StudentExists(student.StudentID)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
             return View(student);
         }
 
-        // GET: Students/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var student = await _context.Students
-                .FirstOrDefaultAsync(m => m.StudentID == id);
-            if (student == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var student = await _context.Students.FirstOrDefaultAsync(m => m.StudentID == id);
+            if (student == null) return NotFound();
             return View(student);
         }
 
-        // POST: Students/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
-            if (student != null)
-            {
-                _context.Students.Remove(student);
-            }
-
+            if (student != null) _context.Students.Remove(student);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool StudentExists(int id)
-        {
-            return _context.Students.Any(e => e.StudentID == id);
-        }
+        private bool StudentExists(int id) => _context.Students.Any(e => e.StudentID == id);
     }
 }
